DSND-3110: Add delta at to delete endpoint
* Add internal to put and delete uris
* Add stale delta check to delete
* Handle resource not found error in delete processing
* Fix broken tests

+++ src/main/java/uk/gov/companieshouse/company/profile/service/CompanyProfileService.java
@@ -12,6 +12,7 @@ import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.data.mongodb.core.MongoTemplate;
@@ -404,27 +405,36 @@ public class CompanyProfileService {
     /**
      * Delete company profile.
      */
-    @Transactional
-    public void deleteCompanyProfile(String contextId,
-            String companyNumber) throws ResourceNotFoundException {
-        VersionedCompanyProfileDocument companyProfileDocument = getCompanyProfileDocument(companyNumber);
-        Data companyProfile = companyProfileDocument.getCompanyProfile();
-        String parentCompanyNumber = companyProfileDocument.getParentCompanyNumber();
-        if (parentCompanyNumber != null && companyProfile.getType().equals("uk-establishment")) {
-            LinkRequest ukEstablishmentLinkRequest =
-                    new LinkRequest(contextId, parentCompanyNumber,
-                            UK_ESTABLISHMENTS_LINK_TYPE,
-                            UK_ESTABLISHMENTS_DELTA_TYPE, Links::getUkEstablishments);
-            checkForDeleteLink(ukEstablishmentLinkRequest);
+    public void deleteCompanyProfile(String contextId, String companyNumber, String deltaAt) {
+        if (StringUtils.isBlank(deltaAt)) {
+            logger.error("deltaAt missing from delete request", DataMapHolder.getLogMap());
+            throw new BadRequestException("deltaAt is null or empty");
         }
+        // Ask someone about this better to use a try catch for the getCompanyProfileDocument call as it's used by other methods too,
+        // or should change the method itself and fix the other methods in this class that use it
+        try {
+            VersionedCompanyProfileDocument companyProfileDocument = getCompanyProfileDocument(companyNumber);
+            Data companyProfile = companyProfileDocument.getCompanyProfile();
+            String parentCompanyNumber = companyProfileDocument.getParentCompanyNumber();
+            if (parentCompanyNumber != null && companyProfile.getType().equals("uk-establishment")) {
+                LinkRequest ukEstablishmentLinkRequest =
+                        new LinkRequest(contextId, parentCompanyNumber,
+                                UK_ESTABLISHMENTS_LINK_TYPE,
+                                UK_ESTABLISHMENTS_DELTA_TYPE, Links::getUkEstablishments);
+                checkForDeleteLink(ukEstablishmentLinkRequest);
+            }
 
-        companyProfileRepository.delete(companyProfileDocument);
-        companyProfileApiService.invokeChsKafkaApiWithDeleteEvent(contextId, companyNumber,
-                companyProfile);
-
-        logger.info(String.format("Company profile is deleted in MongoDb with companyNumber %s",
-                companyNumber), DataMapHolder.getLogMap());
-
+            companyProfileRepository.delete(companyProfileDocument);
+            logger.info(String.format("Company profile is deleted in MongoDb with companyNumber %s",
+                    companyNumber), DataMapHolder.getLogMap());
+            companyProfileApiService.invokeChsKafkaApiWithDeleteEvent(contextId, companyNumber,
+                    companyProfile);
+        } catch (ResourceNotFoundException ex) {
+            logger.info(String.format("Delete for non-existent document in MongoDb with companyNumber %s",
+                    companyNumber), DataMapHolder.getLogMap());
+            companyProfileApiService.invokeChsKafkaApiWithDeleteEvent(contextId, companyNumber,
+                    new Data());
+        }
     }
 
     /**
